+++ R/plosabstract.R
@@ -8,7 +8,6 @@
 #' plosabstract(q = 'drosophila', fl=c('id','author'), limit = 5)
 #' plosabstract(q = 'drosophila', fl='author', limit = 5)
 #' plosabstract(q = 'drosophila', fl=c('id','author','title'), limit = 5)
-#' plosabstract(q = 'drosophila', fl=c('id','author','title'), limit = 5, config=verbose())
 #' }
 
 plosabstract <- function(q = NULL, fl = 'id', fq = NULL, sort = NULL, start = 0,