+++ R/rplos-package.R
@@ -56,6 +56,7 @@ NULL
 #' \itemize{
 #'  \item \code{\link{crossref}}: service no longer provided - 
 #'  see the package \code{rcrossref}
+#'  \item \code{\link{citations}}: service no longer available
 #' }
 #'
 #' @name rplos-defunct